Remove ReflectionExtensions. Add TypeExtensions to NET40 and SL.
