+++ Source/OxyPlot.Tests/Utilities/StringHelperTests.cs
@@ -46,6 +46,23 @@ namespace OxyPlot.Tests
             Assert.AreEqual("Item {0}", StringHelper.CreateValidFormatString("Item {0}"), "Item {0}");
         }
 
+        [Test]
+        public void FormatIntegers()
+        {
+            var items = new[] { 1, 2, 3 };
+            CollectionAssert.AreEqual(new[] { "1", "2", "3" }, items.Format(null, null, CultureInfo.InvariantCulture));
+            CollectionAssert.AreEqual(new[] { "01", "02", "03" }, items.Format("", "00", CultureInfo.InvariantCulture));
+            CollectionAssert.AreEqual(new[] { "Item 1", "Item 2", "Item 3" }, items.Format(null, "Item {0}", CultureInfo.InvariantCulture));
+        }
+
+        [Test]
+        public void FormatStrings()
+        {
+            var items = new[] { "One", "Two", "Three" };
+            CollectionAssert.AreEqual(new[] { "3", "3", "5" }, items.Format("Length", null, CultureInfo.InvariantCulture));
+            CollectionAssert.AreEqual(new[] { "Item One", "Item Two", "Item Three" }, items.Format(null, "Item {0}", CultureInfo.InvariantCulture));
+        }
+
         public class Item
         {
             public string Text { get; set; }
