+++ Source/OxyPlot/Utilities/StringHelper.cs
@@ -10,6 +10,8 @@
 namespace OxyPlot
 {
     using System;
+    using System.Collections;
+    using System.Collections.Generic;
     using System.Reflection;
     using System.Text.RegularExpressions;
 
@@ -85,6 +87,35 @@ namespace OxyPlot
             }
 
             return string.Concat("{0:", input, "}");
+        }
+
+        /// <summary>
+        /// Formats each item in a sequence by the specified format string and property.
+        /// </summary>
+        /// <param name="source">The source target.</param>
+        /// <param name="propertyName">The property name.</param>
+        /// <param name="formatString">The format string. The format argument {0} can be used for the value of the property in each element of the sequence.</param>
+        /// <param name="provider">The format provider.</param>
+        /// <exception cref="System.InvalidOperationException">Could not find property.</exception>
+        public static IEnumerable<string> Format(this IEnumerable source, string propertyName, string formatString, IFormatProvider provider)
+        {
+            var fs = CreateValidFormatString(formatString);
+            if (string.IsNullOrEmpty(propertyName))
+            {
+                foreach (var element in source)
+                {
+                    yield return string.Format(provider, fs, element);
+                }
+            }
+            else
+            {
+                var reflectionPath = new ReflectionPath(propertyName);
+                foreach (var element in source)
+                {
+                    var value = reflectionPath.GetValue(element);
+                    yield return string.Format(provider, fs, value);
+                }
+            }
         }
     }
 }
